pyproject.toml
[project]
name = "langchain-eden-marco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "faiss-cpu>=1.11.0",
  "flask>=3.1.1",
  "langchain>=0.3.25",
  "langchain-community>=0.3.25",
  "langchain-core>=0.3.65",
  "langchain-ollama>=0.3.3",
  "langchain-pinecone>=0.2.8",
  "pydantic>=2.11.5",
  "pypdf>=5.6.0",
  "python-dotenv>=1.1.0",
  "questionary>=2.1.0",
  "types-requests>=2.32.4.20250611",
]

[dependency-groups]
dev = ["mypy>=1.16.0", "ruff>=0.11.13"]

# Ruff configuration
[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
]
ignore = [
  "E501", # line too long, handled by black
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# mypy configuration
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
mypy_path = ["src"]
explicit_package_bases = true
